Added javadocs, moving UserDAOTest.java to dao package.
Also fixed toasts in SignupActivity, created checkUserAndCreate method in PocketMealsRepository.
