+++ app/src/androidTest/java/com/example/pocketmeals/dao/UserDAOTest.java
@@ -1,4 +1,4 @@
-package com.example.pocketmeals;
+package com.example.pocketmeals.dao;
 
 import static org.junit.Assert.*;
