+++ app/src/main/java/com/example/pocketmeals/activity/AdminActivity.java
@@ -16,13 +16,26 @@ import java.util.ArrayList;
 /**
  * @author Andrew Lee
  * created: 8/8/2025
- * Explanation: Activity for all admin actions
+ * Explanation: This activity provides administrative functionalities, allowing an admin user
+ * to manage other user accounts. It displays a list of all users and provides
+ * options to delete a user, and grant or revoke admin privileges.
  */
 public class AdminActivity extends AppCompatActivity {
-
-    private AdminAdapter adapter;
+  private AdminAdapter adapter;
   private PocketMealsRepository repository;
 
+  /**
+   * Called when the activity is first created.
+   * This method initializes the activity's layout using view binding, sets up the
+   * {@link PocketMealsRepository}, and configures the {@link androidx.recyclerview.widget.RecyclerView}
+   * with a {@link LinearLayoutManager} and an {@link AdminAdapter}. It uses an
+   * {@link AdminViewModel} to observe all user accounts and updates the adapter accordingly.
+   * It also sets click listeners for the delete, make admin, and remove admin buttons.
+   *
+   * @param savedInstanceState If the activity is being re-initialized after previously being shut down
+   * then this Bundle contains the data it most recently supplied in {@link #onSaveInstanceState(Bundle)}.
+   * Note: Otherwise it is null.
+   */
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
@@ -45,6 +58,12 @@ public class AdminActivity extends AppCompatActivity {
     binding.removeAdminButton.setOnClickListener(v -> removeAdmin());
   }
 
+  /**
+   * Deletes the currently selected user.
+   * This method retrieves the user selected in the {@link AdminAdapter}. If no user
+   * is selected, it shows a toast message. Otherwise, it calls the repository to
+   * delete the user from the database and displays a confirmation toast.
+   */
   private void deleteUser() {
     User user = adapter.getSelectedUser();
     if (user == null) {
@@ -55,6 +74,12 @@ public class AdminActivity extends AppCompatActivity {
     toastMaker("Deleted user: " + user.getUsername());
   }
 
+  /**
+   * Grants admin privileges to the currently selected user.
+   * This method retrieves the selected user, sets their admin status to `true`,
+   * updates the user in the database via the repository, and shows a confirmation toast.
+   * A toast message is shown if no user is selected.
+   */
   private void makeAdmin() {
     User user = adapter.getSelectedUser();
     if (user == null) {
@@ -66,6 +91,12 @@ public class AdminActivity extends AppCompatActivity {
     toastMaker(user.getUsername() + " is now an admin");
   }
 
+  /**
+   * Revokes admin privileges from the currently selected user.
+   * This method retrieves the selected user, sets their admin status to `false`,
+   * updates the user in the database, and shows a confirmation toast.
+   * A toast message is shown if no user is selected.
+   */
   private void removeAdmin() {
     User user = adapter.getSelectedUser();
     if (user == null) {
@@ -77,10 +108,21 @@ public class AdminActivity extends AppCompatActivity {
     toastMaker(user.getUsername() + " is now a regular user");
   }
 
+  /**
+   * Displays a short toast message to the user.
+   *
+   * @param message The message to be displayed in the toast.
+   */
   private void toastMaker(String message) {
     Toast.makeText(this, message, Toast.LENGTH_SHORT).show();
   }
 
+  /**
+   * A factory method to create an {@link Intent} for starting the {@link AdminActivity}.
+   *
+   * @param context The context of the calling activity.
+   * @return A new {@link Intent} configured to start the {@link AdminActivity}.
+   */
   public static Intent adminIntentFactory(Context context) {
     return new Intent(context, AdminActivity.class);
   }
