+++ app/src/main/java/com/example/pocketmeals/database/dao/UserDAO.java
@@ -40,13 +40,13 @@ public interface UserDAO {
   
   @Update
   void update(User... users);
-  
-  //Test purposes to avoid needing to create a testing utility
-  @Query("SELECT * FROM " + PocketMealsDatabase.USER_TABLE + " ORDER BY username")
-  List<User> getUserList();
 
   @Query("SELECT * FROM " + PocketMealsDatabase.USER_TABLE + " WHERE username == :username")
   User getUserByUserNameSync(String username);
+
+  //Test purposes to avoid needing to create a testing utility
+  @Query("SELECT * FROM " + PocketMealsDatabase.USER_TABLE + " ORDER BY username")
+  List<User> getUserList();
 
   @Query("SELECT * FROM " + PocketMealsDatabase.USER_TABLE + " WHERE id == :userId")
   User getUserByUserIdSync(int userId);
